dom/alarm/test/test_alarm_permitted_app.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Permitted Application for Alarm API</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test">
<script type="application/javascript">

"use strict";

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({"set": [["dom.mozAlarms.enabled", true]]}, function() {
  // mozAlarms is installed on all platforms except Android for the moment.
  if (navigator.appVersion.indexOf("Android") != -1) {
    try {
      todo('mozAlarms' in navigator,
           "mozAlarms is not allowed on Android for now. TODO Bug 863557.");
    } catch (e) {
      todo(!e, "('mozAlarms' in navigator) should not throw exceptions once " +
               "mozAlarms is installed on Android. TODO Bug 863557. " +
               "Caught exception: " + e);
    }
  } else {
    ok('AlarmsManager' in window, "Interface AlarmsManager should exist");
    ok('mozAlarms' in navigator, "navigator.mozAlarms should exist");
    ok(navigator.mozAlarms instanceof AlarmsManager,
       "navigator.mozAlarms should be an instance of AlarmsManager");
  }

  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
